+++ Chapter.Net.WPF.Navigation/INavigationService.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using System.Windows;
 using Chapter.Net.WPF.Navigation.Dialogs;
 using Chapter.Net.WPF.Navigation.MessageBoxes;
+using Chapter.Net.WPF.Navigation.Presenters;
 using Chapter.Net.WPF.Navigation.Windows;
 
 namespace Chapter.Net.WPF.Navigation;