+++ CHANGELOG.md
@@ -6,12 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Moved navigation presenters into an own namespace.
 
 ## [2.0.0] - 2024-06-07
 ### Changed
 - Update to support .Net 8 only.
 ### Fixed
-- Fixed initial assign of the selected path for the folder browser
+- Fixed initial assign of the selected path for the folder browser.
 
 ## [1.2.0] - 2024-03-31
 ### Added